refactor: extract getEventTypeByViewer for private user dashboard event types
