+++ packages/lib/index.ts
@@ -5,3 +5,4 @@ export * from "./isBookingLimits";
 export * from "./isDurationLimits";
 export * from "./validateIntervalLimitOrder";
 export * from "./schedules";
+export * from "./getEventTypeByViewer";
